third_party/WebKit/LayoutTests/inspector/console/console-message-from-inline-with-url-expected.txt

CONSOLE MESSAGE: line 7: foo
CONSOLE ERROR: line 1: Uncaught SyntaxError: Unexpected token }
Tests that log message and syntax errors from inline scripts with sourceURL are logged into console, contains correct link and doesn't cause browser crash.

foo.js:3 foo
boo.js:1 Uncaught SyntaxError: Unexpected token }
    at addInlineWithSyntaxError (foo.js:10)
addInlineWithSyntaxError @ foo.js:10
setTimeout (async)
(anonymous) @ VM:1
